# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sahil Master System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DayRecord {
  date: string;
  fd: number;
  gb: number;
  gl: number;
  ds: number;
}

export interface DayRecordInput {
  date: string;
  fd: number;
  gb: number;
  gl: number;
  ds: number;
}

export interface PendingEntry {
  date: string;
  /** @nullable */
  fd: number | null;
  /** @nullable */
  gb: number | null;
  /** @nullable */
  gl: number | null;
  /** @nullable */
  ds: number | null;
  complete: boolean;
}

export type PendingInputField = typeof PendingInputField[keyof typeof PendingInputField];


export const PendingInputField = {
  fd: 'fd',
  gb: 'gb',
  gl: 'gl',
  ds: 'ds',
} as const;

export interface PendingInput {
  date: string;
  field: PendingInputField;
  value: number;
}

export interface PendingResult {
  saved: boolean;
  complete: boolean;
  entry: PendingEntry;
}

export interface PredictInput {
  fd: number;
  gb: number;
  gl: number;
  ds: number;
  formula: string[];
  gameDate: string;
}

export type PredictOutputTokariItem = {
  jodi: string;
  count: number;
};

export interface PredictOutput {
  tokari: PredictOutputTokariItem[];
  L1: string[];
  L2: string[];
  L3: string[];
  gameDate: string;
}

export interface PredictionEntry {
  gameDate: string;
  L1: string[];
  L2: string[];
  L3: string[];
  status: string;
  matched?: string[];
  actualResult?: string[];
}

export interface CheckInput {
  date: string;
  fd: number;
  gb: number;
  gl: number;
  ds: number;
}

export interface CheckOutput {
  status: string;
  matched: string[];
}

export interface StatsOutput {
  total: number;
  passed: number;
  failed: number;
  pending: number;
  passL1: number;
  passL2: number;
  passL3: number;
  successRate: number;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export interface ErrorResponse {
  error: string;
}