# lib/api-zod/src/generated/types/checkInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sahil Master System API
 * OpenAPI spec version: 0.1.0
 */

export interface CheckInput {
  date: string;
  fd: number;
  gb: number;
  gl: number;
  ds: number;
}