# lib/api-zod/src/generated/types/pendingEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sahil Master System API
 * OpenAPI spec version: 0.1.0
 */

export interface PendingEntry {
  date: string;
  /** @nullable */
  fd: number | null;
  /** @nullable */
  gb: number | null;
  /** @nullable */
  gl: number | null;
  /** @nullable */
  ds: number | null;
  complete: boolean;
}