# lib/api-zod/src/generated/types/pendingResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sahil Master System API
 * OpenAPI spec version: 0.1.0
 */
import type { PendingEntry } from './pendingEntry';

export interface PendingResult {
  saved: boolean;
  complete: boolean;
  entry: PendingEntry;
}