# lib/api-zod/src/generated/types/predictOutput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sahil Master System API
 * OpenAPI spec version: 0.1.0
 */
import type { PredictOutputTokariItem } from './predictOutputTokariItem';

export interface PredictOutput {
  tokari: PredictOutputTokariItem[];
  L1: string[];
  L2: string[];
  L3: string[];
  gameDate: string;
}