# lib/api-zod/src/generated/types/statsOutput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sahil Master System API
 * OpenAPI spec version: 0.1.0
 */

export interface StatsOutput {
  total: number;
  passed: number;
  failed: number;
  pending: number;
  passL1: number;
  passL2: number;
  passL3: number;
  successRate: number;
}